+++ lib/best_in_place.rb
@@ -1,11 +1,9 @@
 require 'rails'
 require 'action_view/railtie'
-require 'active_record/railtie'
-require 'action_controller/model_naming'
+require 'action_controller'
 require 'best_in_place/engine'
-require "best_in_place/utils"
-require "best_in_place/helper"
-require "best_in_place/railtie"
-require "best_in_place/controller_extensions"
-require "best_in_place/display_methods"
-
+require 'best_in_place/utils'
+require 'best_in_place/helper'
+require 'best_in_place/railtie'
+require 'best_in_place/controller_extensions'
+require 'best_in_place/display_methods'