+++ lib/best_in_place/helper.rb
@@ -1,14 +1,12 @@
 module BestInPlace
   module Helper
-
     def best_in_place(object, field, opts = {})
-
       best_in_place_assert_arguments(opts)
       type = opts[:as] || :input
       field = field.to_s
 
-      options={}
-      options[:data]= HashWithIndifferentAccess.new(opts[:data])
+      options = {}
+      options[:data] = HashWithIndifferentAccess.new(opts[:data])
       options[:data]['bip-type'] = type
       options[:data]['bip-attribute'] = field
 
@@ -41,9 +39,7 @@ module BestInPlace
       options[:class] = ['best_in_place'] + Array(opts[:class] || opts[:classes])
       options[:id] = opts[:id] || BestInPlace::Utils.build_best_in_place_id(real_object, field)
 
-
       options[:data]['bip-activator'] = opts[:activator].presence
-
 
       options[:data]['bip-html-attrs'] = opts[:html_attrs].to_json unless opts[:html_attrs].blank?
       options[:data]['bip-inner-class'] = opts[:inner_class].presence
@@ -57,27 +53,23 @@ module BestInPlace
       options[:data]['bip-cancel-button-class'] = opts[:cancel_button_class].presence
       options[:data]['bip-original-content'] = html_escape(value).presence
 
-
       options[:data]['bip-url'] = url_for(opts[:url] || object)
 
       options[:data]['bip-confirm'] = opts[:confirm].presence
       options[:data]['bip-value'] = html_escape(value).presence
 
-
-      if opts[:sanitize].presence.to_s == 'false'
-        options[:data][:sanitize] = false
+      if opts[:raw]
+        options[:data]['bip-raw'] = 'true'
       end
 
-      #delete nil keys only
-      options[:data].delete_if { |k, v| v.nil? }
-
+      # delete nil keys only
+      options[:data].delete_if { |_, v| v.nil? }
       content_tag(:span, options) do
-        !options[:data][:sanitize] ? display_value : display_value.html_safe
+        opts[:raw].present? ? display_value.to_s.html_safe : display_value
       end
-
     end
 
-    def best_in_place_if(condition, object, field, opts={})
+    def best_in_place_if(condition, object, field, opts = {})
       if condition
         best_in_place(object, field, opts)
       else
@@ -85,16 +77,14 @@ module BestInPlace
       end
     end
 
-    def best_in_place_unless(condition, object, field, opts={})
+    def best_in_place_unless(condition, object, field, opts = {})
       best_in_place_if(!condition, object, field, opts)
     end
-
 
     private
 
     def best_in_place_build_value_for(object, field, opts)
       klass = object.class
-
       if opts[:display_as]
         BestInPlace::DisplayMethods.add_model_method(klass, field, opts[:display_as])
         object.send(opts[:display_as]).to_s
@@ -113,6 +103,8 @@ module BestInPlace
           if field_value.blank?
             ''
           else
+
+
             BestInPlace::ViewHelpers.send(opts[:display_with], field_value)
           end
         end
@@ -130,7 +122,7 @@ module BestInPlace
       args.assert_valid_keys(:id, :type, :nil, :class, :collection, :data,
                              :activator, :cancel_button, :cancel_button_class, :html_attrs, :inner_class, :nil,
                              :object_name, :ok_button, :ok_button_class, :display_as, :display_with, :path,
-                             :use_confirm, :confirm, :sanitize, :helper_options, :url, :place_holder, :class, :as, :param)
+                             :use_confirm, :confirm, :sanitize, :raw, :helper_options, :url, :place_holder, :class, :as, :param)
 
       best_in_place_deprecated_options(args)
 
@@ -174,12 +166,16 @@ module BestInPlace
         ActiveSupport::Deprecation.warn('[Best_in_place] :use_confirm is deprecated in favor of :confirm ')
       end
 
+      if deprecated_option = args.delete(:sanitize)
+        args[:raw] = !deprecated_option
+        ActiveSupport::Deprecation.warn('[Best_in_place] :sanitize is deprecated in favor of :raw ')
+      end
     end
 
     def best_in_place_collection_builder(collection)
       case collection
         when Array
-          Hash[collection.collect { |x| [x.to_s, x.to_s] }]
+          Hash[collection.map { |x| [x.to_s, x.to_s] }]
         else
           collection.stringify_keys
       end
@@ -191,6 +187,3 @@ module BestInPlace
     end
   end
 end
-
-
-