+++ lib/best_in_place/railtie.rb
@@ -1,7 +1,8 @@
-require 'jquery/rails/railtie'
+require 'rails/railtie'
+require 'action_view/base'
 
 module BestInPlace
-  class Railtie < ::Rails::Railtie
+  class Railtie < ::Rails::Railtie #:nodoc:
     config.after_initialize do
       BestInPlace::ViewHelpers = ActionView::Base.new
     end