+++ lib/best_in_place/utils.rb
@@ -1,8 +1,6 @@
 module BestInPlace
-  module Utils
+  module Utils #:nodoc:
     module_function
-    extend ActionController::ModelNaming
-
     def build_best_in_place_id(object, field)
       case object
         when Symbol, String
@@ -18,5 +16,14 @@ module BestInPlace
     def object_to_key(object)
       model_name_from_record_or_class(object).param_key
     end
+
+    def convert_to_model(object)
+      object.respond_to?(:to_model) ? object.to_model : object
+    end
+
+    def model_name_from_record_or_class(record_or_class)
+      (record_or_class.is_a?(Class) ? record_or_class : convert_to_model(record_or_class).class).model_name
+    end
+
   end
 end