+++ spec/internal/app/controllers/admin/users_controller.rb
@@ -7,8 +7,14 @@ class Admin::UsersController < ApplicationController
     @user = User.find(params[:id])
 
     respond_to do |format|
-      @user.update_attributes(params[:user])
+      @user.update_attributes(user_params)
       format.json { respond_with_bip(@user) }
     end
+  end
+
+  private
+
+  def user_params
+    params.require(:user).permit!
   end
 end